third_party/WebKit/LayoutTests/fast/canvas-api/canvas-filter-value.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(t) {
    var canvas = document.createElement('canvas');
    var ctx = canvas.getContext('2d');
    
    assert_equals(ctx.filter, 'none');
    ctx.filter = 'blur(5px)';
    assert_equals(ctx.filter, 'blur(5px)');
    
    ctx.save();
    ctx.filter = 'none';
    assert_equals(ctx.filter, 'none');
    ctx.restore();
    assert_equals(ctx.filter, 'blur(5px)');
    
    // Invalid filter should be ignored
    ctx.filter = 'blur(10)';
    assert_equals(ctx.filter, 'blur(5px)');
    ctx.filter = 'blur 10px';
    assert_equals(ctx.filter, 'blur(5px)');
    
    ctx.filter = 'inherit';
    assert_equals(ctx.filter, 'blur(5px)');
    ctx.filter = 'initial';
    assert_equals(ctx.filter, 'blur(5px)');
    ctx.filter = 'unset';
    assert_equals(ctx.filter, 'blur(5px)');
    
    ctx.filter = '';
    assert_equals(ctx.filter, 'blur(5px)');
    ctx.filter = null;
    assert_equals(ctx.filter, 'blur(5px)');
    ctx.filter = undefined;
    assert_equals(ctx.filter, 'blur(5px)');
    
    // verify that exact string is preserved
    ctx.filter = 'blur(  5px)';
    assert_equals(ctx.filter, 'blur(  5px)');
}, 'Test that createImageBitmap from a bitmaprenderer canvas produces correct result');
</script>
